% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebuilding_helper_functions.R
\name{calc.pyclone.ci}
\alias{calc.pyclone.ci}
\title{Function to compute confidence intervals of each cluster in each region}
\usage{
calc.pyclone.ci(pyclone, pyclust, nclusters, prefix = "LTX", lower_min = 0)
}
\arguments{
\item{pyclone}{An R list object containing information about the PhyloCCF
of each mutation in each tumour region.}
}
\value{
An R list containing elements: 'ccf_cluster_table', 'mean_phylo_ccf',
'median_pyclone_ccf', 'median_phylo_ccf', 'ccf_ci_upper', 'ccf_ci_lower',
'ccf_ci_boot_upper', 'ccf_ci_boot_lower'
}
\description{
This function takes in an R list containing mutation PhyloCCF in each region,
and mutation assignments to a cluster and computes bootstrapped
confidence intervals.
}
