% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.building.functions.R
\name{clusterDistributionAcrossGenome}
\alias{clusterDistributionAcrossGenome}
\title{Function to determine the distribution of location of mutations within a cluster}
\usage{
clusterDistributionAcrossGenome(
  cluster,
  clonal_cluster,
  test.pyclone,
  iterations = 10000
)
}
\description{
Function to determine the distribution of location of mutations within a cluster
}
