% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_clustering_functions.R
\name{clustering_preprocess}
\alias{clustering_preprocess}
\title{Input data preprocessing function}
\usage{
clustering_preprocess(
  input_table,
  new.dir,
  subclonal_copy_correction = TRUE,
  multiple_test_correction = TRUE,
  only_truncal_subclonal_copy_correction = TRUE,
  fix_absentCCFs = TRUE
)
}
\arguments{
\item{input_table}{An dataframe of the input mutation table in the correct
format. For more information on the input table format, please see our
tree building protocol.}

\item{new.dir}{A character specifying the directory where the pyclone
output should be saved.}

\item{subclonal_copy_correction}{A logical value that specifies whether subclonal
copy number correction should be performed. 
Default is set to TRUE}

\item{multiple_test_correction}{A logical value that specifies whether multiple 
testing correction should be applied for the copy number correcting mutations.
Default is set to TRUE}

\item{only_truncal_subclonal_copy_correction}{A logical value that specifies 
whether only truncal subclonal copy number correction should be used.
Default is set to TRUE}

\item{fix_absentCCFs}{A logical value that specifies whether CCF 
of absent mutations should be set to zero.
Default is set to TRUE}
}
\value{
list including patient, regions.to.use, mut.table, seg.mat.copy
seg.mat.phylo, phylo.region.list, simpleClusterList
}
\description{
This function takes the input tsv and formats the data to be compatible with
the main CONIPHER clustering function.
}
