% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_clustering_functions.R
\name{clustering_run}
\alias{clustering_run}
\title{Main clustering function}
\usage{
clustering_run(
  input_list,
  nProcs,
  new.dir,
  burn_in,
  pyclone_seed,
  template.config.yaml
)
}
\arguments{
\item{input_list}{A list created by the clustering preprocess function
including patient id, regions to use, phylo region list and others.}

\item{nProcs}{A value referring to how many parallel processes 
of pyclone should be run.}

\item{new.dir}{A character specifying the directory where the pyclone
output should be saved.}

\item{burn_in}{Burn-in for DP clustering.}

\item{pyclone_seed}{Seed for PyClone run.}

\item{template.config.yaml}{Location of the template yaml file used to run PyClone.}
}
\value{
sample.results which is the location of the pyclone output table.
}
\description{
This function takes the input list created in the preprocessing along with
the number of cores and output directory to run the main clustering.
}
