% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.building.functions.R
\name{color.tree}
\alias{color.tree}
\title{Plotting function to colour the nodes on the phylogenetic tree}
\usage{
color.tree(edgelength, opacity = 255)
}
\arguments{
\item{edgelength}{A named vector containing number of mutations of each cluster}
}
\description{
Plotting function to colour the nodes on the phylogenetic tree
}
