% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebuilding_helper_functions.R
\name{compute_subclonal_expansion_score}
\alias{compute_subclonal_expansion_score}
\title{Function to compute subclonal expansion score on a selected alternative tree,
on a tumour sample and whole tumour level.
The subclonal expansion score for each tumour sample is computed as the
maximum CCF of any of the terminal (leaf) nodes present in that tumour sample.
Note, for multi-sample cases, there may exist a sample with no terminal nodes
present, in which case the subclonal expansion score for this sample is set
to 0. The tumour level subclonal expansion score is taken as the maximum
subclonal expansion score across tumour samples.}
\usage{
compute_subclonal_expansion_score(tree_list, tree_id, ccf_table_pyclone_clean)
}
\arguments{
\item{tree_list}{A list of tree matrices}

\item{tree_id}{The tree index of the selected alternative tree for which you
want to compute the subclonal expansion score}

\item{ccf_table_pyclone_clean}{The output mutation PhyloCCF data frame that is
computed as part of CONIPHER tree building}
}
\value{
subclonal_exp_score_df, a data frame with the subclonal expansion
score computed for each tumour sample (column subclonal_expansion_score),
and across the whole tumour (column subclonal_expansion_score_tumour).
each clone (rows) in each tumour sample (columns)
}
\description{
Function to compute subclonal expansion score on a selected alternative tree,
on a tumour sample and whole tumour level.
The subclonal expansion score for each tumour sample is computed as the
maximum CCF of any of the terminal (leaf) nodes present in that tumour sample.
Note, for multi-sample cases, there may exist a sample with no terminal nodes
present, in which case the subclonal expansion score for this sample is set
to 0. The tumour level subclonal expansion score is taken as the maximum
subclonal expansion score across tumour samples.
}
