% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_subclone_proportions.R
\name{compute_subclone_proportions}
\alias{compute_subclone_proportions}
\title{Function to compute clone proportions on a selected alternative tree structure}
\usage{
compute_subclone_proportions(
  tree_list,
  ccf_cluster_table,
  clonality_table,
  trunk,
  force_clonal_100 = TRUE,
  tree_id = 1
)
}
\arguments{
\item{tree_list}{A list of tree matrices}

\item{ccf_cluster_table}{A matrix of mean PhyloCCF of each cluster in
each tumour region}

\item{clonality_table}{A matrix of clonality calls for each cluster in
each tumour region}

\item{trunk}{The name of the truncal cluster}

\item{force_clonal_100}{A logical indicating whether to for clusters that are
'clonal' in a region to have CCF==100}

\item{tree_id}{The tree index of the selected alternative tree for which you
want to compute the clone proportions}
}
\description{
Function to compute clone proportions on a selected alternative tree structure
}
