% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.building.functions.R
\name{compute_sum_condition_error}
\alias{compute_sum_condition_error}
\title{Function to compute the Sum Condition Error for a list of trees}
\usage{
compute_sum_condition_error(tree_list, ccf_cluster_table, trunk)
}
\arguments{
\item{tree_list}{A list of tree matrices}

\item{ccf_cluster_table}{A matrix of mean PhyloCCF of each cluster in
each tumour region}

\item{trunk}{The name of the truncal cluster}
}
\value{
sce_vec, A named vector of the sum condition error (SCE) for each
tree structure in the input tree list
}
\description{
This function takes a list of phylogenetic tree structures and the PhyloCCF cluster table and
computes the sum condition error for each tree.
}
