% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.building.functions.R
\name{compute_tree_edge_probability}
\alias{compute_tree_edge_probability}
\title{Function to compute the edge probability score for a list of trees}
\usage{
compute_tree_edge_probability(tree_list, edgelength, trunk)
}
\arguments{
\item{tree_list}{A list of tree matrices}

\item{edgelength}{A named vector containing number of mutations of each cluster}

\item{trunk}{The name of the truncal cluster}
}
\value{
sce_vec, A named vector of the sum condition error (SCE) for each
tree structure in the input tree list
}
\description{
This function takes a list of phylogenetic tree structures and the PhyloCCF cluster table and
computes the sum condition error for each tree.
}
