% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_clustering_functions.R
\name{conipher_clustering}
\alias{conipher_clustering}
\title{Full clustering run function}
\usage{
conipher_clustering(
  case_id,
  out_dir,
  input_tsv_loc,
  input_seg_tsv_loc = NULL,
  subclonal_copy_correction = TRUE,
  only_truncal_subclonal_copy_correction = TRUE,
  pyclone_yaml_loc = NULL,
  min_cluster_size = 5,
  multiple_test_correction = TRUE,
  clean_clusters = TRUE,
  clonal_cutOff = 0.9,
  propClonal_threshold = 0.25,
  fix_absentCCFs = TRUE,
  driver_filter = "1A,1,2A",
  burn_in = 1000,
  seed = 1024,
  nProcs = 1,
  ...
)
}
\arguments{
\item{opt}{a list of options}
}
\description{
This function takes all the input options and runs the three main steps:
preprocess, clustering run and postprocess
}
