% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_conipher_run.R
\name{conipher_run}
\alias{conipher_run}
\title{Full CONIPHER run}
\usage{
conipher_run(case_id, prefix, out_dir, input_tsv_loc, ...)
}
\arguments{
\item{opt}{a list of options}
}
\description{
This function takes all the input options and runs the three main steps:
preprocess, tree building run and postprocess
}
