% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_treebuilding_functions.R
\name{conipher_treebuilding}
\alias{conipher_treebuilding}
\title{Full tree building run function}
\usage{
conipher_treebuilding(
  input_tsv_loc,
  out_dir,
  prefix,
  ccf_buffer = 10,
  pval_cutoff = 0.01,
  use_boot = TRUE,
  merge_clusters = TRUE,
  correct_cpn_clusters = TRUE,
  adjust_noisy_clusters = FALSE,
  adjust_noisy_clusters_prop = 0.05,
  min_ccf = 0.01,
  min_cluster_size = 5,
  multi_trees = TRUE,
  ...
)
}
\arguments{
\item{opt}{a list of options}
}
\description{
This function takes all the input options and runs the three main steps:
preprocess, tree building run and postprocess
}
