% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.building.functions.R
\name{correct.clonality.nesting}
\alias{correct.clonality.nesting}
\title{Function to correct nesting based on cluster clonality}
\usage{
correct.clonality.nesting(
  nestedlist,
  pyclone,
  clonality_table,
  pval_cutoff = 0.01
)
}
\description{
Function to correct nesting based on cluster clonality
}
