% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.building.functions.R
\name{createAllPathsList}
\alias{createAllPathsList}
\title{Function to create a list of all tree paths}
\usage{
createAllPathsList(tree.structure, trunk)
}
\arguments{
\item{pyclone}{An R matrix describing the tree structure with two columns
specifying 'parent' (column 1) and child (column 2)}
}
\value{
An R list of all tree paths from trunk to leaves
}
\description{
Function to create a list of all tree paths
}
