% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebuilding_helper_functions.R
\name{determine.cluster.nesting}
\alias{determine.cluster.nesting}
\title{Function to determine cluster nesting structure.}
\usage{
determine.cluster.nesting(
  pyclone,
  prefix = "LTX",
  min_cluster_size = 5,
  pval_cutoff = 0.01,
  min_ccf = 0.01,
  use_boot = FALSE
)
}
\arguments{
\item{pyclone}{An R list object containing information about the PhyloCCF
of each mutation in each tumour region.}

\item{prefix}{A character string indicating the sample and tumour case prefix}

\item{min_cluster_size}{Threshold for minimum number of mutations required in
a mutation cluster}

\item{pval_cutoff}{A p-value significance threshold for testing whether
clusters can be nested. (i.e. a p-value < pval_cutoff is significant)}

\item{min_ccf}{Minimum threshold for cluster PhyloCCF allowed to be classified
as present}

\item{use_boot}{Whether to use bootstrapping to determine confidence intervals
for each mutation cluster}
}
\value{
'nestedlist', an R list containing information about the nesting
structure of mutation clusters in each region. Elements of the list include:
'nestedclust',  'ccf_ci_lower', 'ccf_ci_upper', 'ccf_cluster_table', 'cluster_qc'.
}
\description{
This function takes as input an R list containing information about PhyloCCF
of each mutation (pyclone) and computes confidence intervals for each mutation
cluster, of the PhyloCCF distributions of the mutations in that cluster. If
use_boot==TRUE, then confidence intervals are computed using bootstrapping.
The function then performs a statistical test (Wilcoxon) for every pair of
clusters to determine whether one cluster can be nested within another.
}
