% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.building.functions.R
\name{extract_consensus_relationships}
\alias{extract_consensus_relationships}
\title{Function to take a list of trees and identify the consensus relationships}
\usage{
extract_consensus_relationships(tree_list, output_as_table = FALSE)
}
\description{
Function to take a list of trees and identify the consensus relationships
}
