% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebuilding_helper_functions.R
\name{extract_daughters}
\alias{extract_daughters}
\title{Function to extract all daughter clones from a parent using a phylogenetic tree}
\usage{
extract_daughters(tree, parent.clones)
}
\arguments{
\item{tree}{A phylogenetic tree matrix with two columns specifying
'parent' (column 1) and child (column 2)}

\item{parent.clones}{The name of the parent clone(s) for which you wish to find all daughters}
}
\description{
Function to extract all daughter clones from a parent using a phylogenetic tree
}
