% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebuilding_helper_functions.R
\name{get_terminal_clusters}
\alias{get_terminal_clusters}
\title{Function to extract the terminal nodes of a phylogenetic tree}
\usage{
get_terminal_clusters(tree_structure)
}
\arguments{
\item{tree_structure}{A matrix of a tree structure (edge matrix)}
}
\value{
A vector of the terminal nodes in the tree
}
\description{
Function to extract the terminal nodes of a phylogenetic tree
}
