% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.building.functions.R
\name{get_tree_level}
\alias{get_tree_level}
\title{Function to compute the tree level of a cluster}
\usage{
get_tree_level(tree_graph, cluster)
}
\arguments{
\item{tree_graph}{A matrix of a tree structure}

\item{cluster}{Name of a cluster for which you want to get the tree level}
}
\description{
Function to compute the tree level of a cluster
}
