% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebuilding_helper_functions.R
\name{grow.multi.trees}
\alias{grow.multi.trees}
\title{Function to determine all possible alternative phylogenies}
\usage{
grow.multi.trees(nestedlist, graph_pyclone, pyclone, ccf_buffer = 10)
}
\arguments{
\item{nestedlist}{An R list containing information about the nesting
structure of mutation clusters in each region.}

\item{graph_pyclone}{An R list containing information about the tree structure}

\item{pyclone}{An R list containing information about mutation PhyloCCF}

\item{ccf_buffer}{PhyloCCF buffer permitted when checking tree level issue}
}
\value{
An R list containing all possible alternative tree structures and
information about which branches are consensus across multiple trees
}
\description{
Function to determine all possible alternative phylogenies
}
