% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebuilding_helper_functions.R
\name{grow.trees}
\alias{grow.trees}
\title{Function to determine default tree structure}
\usage{
grow.trees(
  nestedlist,
  pyclone,
  min_cluster_size = 5,
  ccf_buffer = 10,
  force_trunk = TRUE,
  skip_size = 20
)
}
\arguments{
\item{nestedlist}{An R list containing information about the nesting
structure of mutation clusters in each region.}

\item{pyclone}{An R list object containing information about the PhyloCCF
of each mutation in each tumour region.}

\item{min_cluster_size}{Threshold for minimum number of mutations required in
a mutation cluster}

\item{ccf_buffer}{PhyloCCF buffer permitted when checking tree level issue}

\item{prefix}{A character string indicating the sample and tumour case prefix}
}
\description{
This function takes as input an R list containing information about
PhyloCCF of each mutation (pyclone) and an R list containing a cluster ccf table
and nesting matrix (nestedlist), and returns an R list containing the default tree
structure.
}
