% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebuilding_helper_functions.R
\name{is.there.ccf.issue}
\alias{is.there.ccf.issue}
\title{Function to check whether there is a CCF level issue}
\usage{
is.there.ccf.issue(
  nestedclust,
  directed_input_graph,
  ccf_ci_lower,
  trunk_cluster,
  clusters_to_remove,
  clusters_to_use,
  max_per_level = 115
)
}
\description{
Function to check whether there is a CCF level issue
}
