% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebuilding_helper_functions.R
\name{permute.clusters.to.remove}
\alias{permute.clusters.to.remove}
\title{Function to test which clusters are best to remove from tree
This function permutes through clusters and checks which are the best to remove
from the tree - takes into account number of mutations}
\usage{
permute.clusters.to.remove(
  test_out,
  nestedclust,
  max_per_level,
  tlevels,
  trunk_cluster,
  cluster_qc,
  ccf_ci_lower,
  nclusters,
  speed_cluster = 15
)
}
\description{
Function to test which clusters are best to remove from tree
This function permutes through clusters and checks which are the best to remove
from the tree - takes into account number of mutations
}
