% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebuilding_helper_functions.R
\name{process_mean_cluster_ccfs}
\alias{process_mean_cluster_ccfs}
\title{Function to extract the terminal nodes of a phylogenetic tree}
\usage{
process_mean_cluster_ccfs(ccf_table_pyclone_clean)
}
\arguments{
\item{ccf_table_pyclone_clean}{A mutation table with PhyloCCF values. This
table is an item in the R list object list sample_pyclone_tree, which is the
output of the CONIPHER treebuilding_run() function.}
}
\description{
Function to extract the terminal nodes of a phylogenetic tree
}
