% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.building.functions.R
\name{remove_clustered_clones}
\alias{remove_clustered_clones}
\title{Function to remove mutation clusters with genomically clustered mutations}
\usage{
remove_clustered_clones(
  test.pyclone,
  clonal_cluster,
  p_value_cut = 0.01,
  clustering_estimate_cut = 2
)
}
\description{
Function to remove mutation clusters with genomically clustered mutations
}
