% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebuilding_helper_functions.R
\name{test.distributions}
\alias{test.distributions}
\title{Test difference in PhyloCCF distributions of each pair of mutation clusters}
\usage{
test.distributions(pyclone, nclusters, pval_cutoff = 0.05)
}
\arguments{
\item{pyclone}{An R list object containing information about the PhyloCCF
of each mutation in each tumour region.}

\item{nclusters}{Number of clusters}

\item{pval_cutoff}{A p-value significance threshold for testing whether
clusters can be nested. (i.e. a p-value < pval_cutoff is significant)}
}
\value{
This function returns list of nesting matrices. Each element of the
list is a nesting matrix for one tumour region, that describes whether a
cluster A (row) can be nested within a cluster B (column).
}
\description{
This function compares the distributions of the PhyloCCF of each pair of
mutation clusters in the dataset and outputs a nesting matrix.
}
