% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracerx.tree.function.R
\name{tree.plotting.function}
\alias{tree.plotting.function}
\title{TRACERx tree plotting function}
\usage{
tree.plotting.function(sample_pyclone_tree)
}
\arguments{
\item{sample_pyclone_tree}{A list containing all information about the
tree inferred using function tracerx.tree.building()}
}
\description{
This function is the CONIPHER function to plot the inferred phylogenetic tree.
}
