% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_treebuilding_functions.R
\name{treebuilding_plot}
\alias{treebuilding_plot}
\title{TRACERx tree plotting function}
\usage{
treebuilding_plot(sample_pyclone_tree)
}
\arguments{
\item{sample_pyclone_tree}{A list containing all information about the
tree inferred using function tracerx.tree.building()}
}
\description{
This function is the CONIPHER function to plot the inferred phylogenetic tree.
}
