% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_treebuilding_functions.R
\name{treebuilding_preprocess}
\alias{treebuilding_preprocess}
\title{Input data preprocessing function}
\usage{
treebuilding_preprocess(input_table, prefix, out_dir)
}
\arguments{
\item{input_table}{An dataframe of the input mutation table in the correct
format. For more information on the input table format, please see our
tree building protocol.}

\item{prefix}{A tumour case and sample prefix, e.g. 'CRUK'.}

\item{out_dir}{A file path to the desired output directory}
}
\description{
This function takes the input tsv and formats the data to be compatible with
the main CONIPHER tree building function. NOTE: it is assumed that
clustering has been carried out prior to running tree building.
}
