% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_treebuilding_functions.R
\name{treebuilding_run}
\alias{treebuilding_run}
\title{TRACERx tree building function}
\usage{
treebuilding_run(
  sample_input_list,
  ccf_buffer = 10,
  pval_cutoff = 0.01,
  use_boot = TRUE,
  merge_clusters = TRUE,
  correct_cpn_clusters = TRUE,
  adjust_noisy_clusters = FALSE,
  adjust_noisy_clusters_prop = 0.05,
  min_ccf = 0.01,
  min_cluster_size = 5,
  run.multi.trees = TRUE,
  n_clusters_to_move = 5
)
}
\arguments{
\item{sample_input_list}{An R list object. This object contains information
about each mutation in each tumour region sampled, including which cluster
each mutation was assigned to in mutation clustering.
This object can be generated by running the data preprocessing function}

\item{ccf_buffer}{PhyloCCF buffer allowance for testing tree level issue
(default=10)}

\item{pval_cutoff}{P-value cut off for testing cluster nesting (default=0.01)}

\item{use_boot}{Should bootstrapping be used to compute confidence interval?
(default=TRUE)}

\item{merge_clusters}{Should similar clusters be merged if possible?
(default=TRUE)}

\item{correct_cpn_clusters}{Should clusters driven by copy number errors be
removed? (default=TRUE)}

\item{adjust_noisy_clusters}{Should noisy clusters be adjusted? (default=TRUE)}

\item{adjust_noisy_clusters_prop}{What is the minimum proportion of mutations
required to be present in a region to avoid cluster adjustment? (default=0.05)}

\item{min_ccf}{What is the minimum CCF threshold to consider a mutation as
present? (default=0.01)}

\item{min_cluster_size}{What is the minimum number of mutations required in a
cluster to be included in analysis? (default=5)}

\item{run.multi.trees}{Should alternative tumour phylogenies be explored?
(default=TRUE)}

\item{n_clusters_to_move}{When running multiple trees specify the maximum
number of clusters to attempt moving. (default=5)}
}
\value{
sample_pyclone_tree, an R list object containing output information
from CONIPHER tree building
}
\description{
This function is the main CONIPHER wrapper function to run phylogenetic
tree building from mutation clustering output. NOTE: it is assumed that
clustering has been carried out prior to running tree building.
}
