% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebuilding_helper_functions.R
\name{correct.clonality.nesting}
\alias{correct.clonality.nesting}
\title{Function to correct nesting based on cluster clonality}
\usage{
correct.clonality.nesting(
  nestedlist,
  pyclone,
  clonality_table,
  pval_cutoff = 0.01,
  min_cluster_size = 5,
  use_boot = TRUE,
  min_ccf = 0.05,
  prefix = prefix
)
}
\description{
Function to correct nesting based on cluster clonality
}
