% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.building.functions.R
\name{determine.cluster.nesting}
\alias{determine.cluster.nesting}
\title{Function to determine cluster nesting structure. The function determines for
every pair of clusters whether one cluster can be nested within another.}
\usage{
determine.cluster.nesting(
  pyclone,
  prefix = "LTX",
  min_cluster_size = 5,
  pval_cutoff = 0.01,
  min_ccf = 0.01,
  use_boot = FALSE
)
}
\arguments{
\item{pyclone}{An R list object comprising input data.
This object can be generated by running the tracerx.tree.building function}

\item{prefix}{A character string indicating the sample and tumour case prefix}

\item{min_cluster_size}{Threshold for minimum number of mutations required in
a mutation cluster}

\item{pval_cutoff}{A p-value significance threshold for testing whether
clusters can be nested. (i.e. a p-value < pval_cutoff is significant)}

\item{min_ccf}{Minimum threshold for cluster PhyloCCF allowed to be classified
as present}

\item{use_boot}{Whether to use bootstrapping to determine confidence intervals
for each mutation cluster}
}
\description{
This function takes in the following parameters:
}
